requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
numpy==2.1.3
pillow==11.3.0
pydantic==2.11.7
python-multipart==0.0.20
tensorflow-cpu==2.19.0
